32 bit int for LiDisplay wh tracking

+++ Firmware/firmwareLiBCM/src/energy.cpp
@@ -5,24 +5,24 @@
 
 #include "libcm.h"
 
-uint16_t wattHours_assist = 0;
-uint16_t wattHours_regen = 0;
-uint16_t wattHours_gridCharger = 0;
+uint32_t wattHours_assist = 0;
+uint32_t wattHours_regen = 0;
+uint32_t wattHours_gridCharger = 0;
 
 // Users with LiDisplay will have their own trip meter they can reset, most likely at each fill up.
-static uint16_t tripmeter_wattHours_assist = 0;
-static uint16_t tripMeter_wattHours_regen = 0;
-static uint16_t tripMeter_wattHours_gridCharger = 0;
+static uint32_t tripmeter_wattHours_assist = 0;
+static uint32_t tripMeter_wattHours_regen = 0;
+static uint32_t tripMeter_wattHours_gridCharger = 0;
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-uint16_t energy_getAssist_Wh(void) { return wattHours_assist; }
-uint16_t energy_getRegen_Wh (void) { return wattHours_regen;  }
-uint16_t energy_getGridCharger_Wh (void) { return wattHours_gridCharger;  }
+uint32_t energy_getAssist_Wh(void) { return wattHours_assist; }
+uint32_t energy_getRegen_Wh (void) { return wattHours_regen;  }
+uint32_t energy_getGridCharger_Wh (void) { return wattHours_gridCharger;  }
 
-uint16_t energy_getTripMeterAssist_Wh(void) { return tripmeter_wattHours_assist; }
-uint16_t energy_getTripMeterRegen_Wh (void) { return tripMeter_wattHours_regen;  }
-uint16_t energy_getTripMeterGridCharge_Wh(void) { return tripMeter_wattHours_gridCharger; }
+uint32_t energy_getTripMeterAssist_Wh(void) { return tripmeter_wattHours_assist; }
+uint32_t energy_getTripMeterRegen_Wh (void) { return tripMeter_wattHours_regen;  }
+uint32_t energy_getTripMeterGridCharge_Wh(void) { return tripMeter_wattHours_gridCharger; }
 
 /////////////////////////////////////////////////////////////////////////////////////////
